requirements_gui.txt
yt-dlp>=2023.03.04
tqdm>=4.65.0
customtkinter>=5.2.0
pillow>=10.0.0
